pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quidclaw"
version = "0.3.0"
description = "Zero-barrier personal CFO — AI-powered, local-first, powered by Beancount"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-2.0-only"
authors = [
    { name = "Yue Jiang" },
]
keywords = ["finance", "accounting", "beancount", "ai-agent", "personal-cfo", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial :: Accounting",
]
dependencies = [
    "beancount>=3.0.0",
    "beanquery>=0.1.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/thorb/quidclaw"
Repository = "https://github.com/thorb/quidclaw"
Issues = "https://github.com/thorb/quidclaw/issues"

[project.optional-dependencies]
prices = ["beanprice>=0.1.0"]
dev = [
    "pytest>=8.0",
]

[project.scripts]
quidclaw = "quidclaw.cli:main"

[tool.pytest.ini_options]
markers = [
    "e2e: End-to-end tests that run AI CLI (slow, costs API)",
]
